Organize feature libraries in subdirectory
